refactor: [#65] controller 분리
