+++ src/main/java/com/techeeresc/tab/domain/post/controller/PostDeleteMethodController.java
@@ -1,0 +1,56 @@
+package com.techeeresc.tab.domain.post.controller;
+
+import com.techeeresc.tab.domain.post.dto.mapper.PostMapper;
+import com.techeeresc.tab.domain.post.service.PostService;
+import com.techeeresc.tab.global.exception.response.ErrorResponse;
+import com.techeeresc.tab.global.status.StatusMessage;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@ApiResponses({
+  @ApiResponse(
+      responseCode = "400",
+      description = "BAD REQUEST by Parameter Missing",
+      content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+  @ApiResponse(
+      responseCode = "400",
+      description = "BAD REQUEST by Type Mismatch",
+      content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+  @ApiResponse(
+      responseCode = "404",
+      description = "NOT FOUND",
+      content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+  @ApiResponse(
+      responseCode = "500",
+      description = "INTERNAL SERVER ERROR",
+      content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+})
+@Tag(name = "post", description = "Post API")
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/post")
+public class PostDeleteMethodController {
+  private final PostService POST_SERVICE;
+  private final PostMapper POST_MAPPER;
+
+  @Operation(
+      summary = "delete post",
+      description = "Method: DELETE, success response code: 200, 처리 완료 후 메인 화면 이동 필요")
+  @ApiResponse(responseCode = "200", description = "OK")
+  @DeleteMapping("/{id}")
+  public ResponseEntity<String> deletePost(
+      @Parameter(description = "페이지 아이디", in = ParameterIn.PATH) @PathVariable Long id) {
+    POST_SERVICE.deletePost(id);
+    return new ResponseEntity<>(StatusMessage.OK.getStatusMessage(), HttpStatus.OK);
+  }
+}
