+++ src/main/java/com/techeeresc/tab/domain/post/repository/PostQueryDslRepository.java
@@ -7,6 +7,5 @@ import java.util.List;
 
 public interface PostQueryDslRepository {
   List<Post> findByTitleContainsWordWithQueryDsl(String word, Pageable pageable);
-
   List<Post> findAllPostListWithQueryDsl(Pageable pageable);
 }
